refactor(test): 테스트 클래스에 이름 붙이고 Nested 사용해 역할 별로 테스트 분류
